+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`1.12.5...main`][1.12.5...main].
 
+### Enhanced
+
+- Adjusted `Rules\Files\ReferenceNamespacedSymbolsRelativeToNamespacePrefixRector` to allow configuring parent namespace prefixes for automatic discovery of namespace prefixes ([#310]), by [@localheinz]
+
 ## [`1.12.5`][1.12.5]
 
 For a full diff see [`1.12.4...1.12.5`][1.12.4...1.12.5].